+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to semantic versioning.
 
 ## _Unreleased_
 
+- Improve `fromString` in `URLPath` to infallibly parse all strings.
+- Remove `fromStringO` from `URLPath`.
 - Fix `toURL`/`toURLO` in `URLPath` failing to prioritise the retention of the provided base URL.
 
 ## 0.19.1 (2024-04-08)